expose a HealthCheckStatsListener hook to query the stats cache
As part of debugging problems in the cross-cell region support,
it was important to know whether or not a tablet was being populated
in the TabletStatsCache, but the existing debug status UI didn't
expose this information.

To make this work, add a new method to the HealthCheckStatsListener
to query whether or not a given tablet is being tracked in the
downstream cache, and call this from the debug UI when populating
the healthcheck data.

Signed-off-by: Michael Demmer <mdemmer@slack-corp.com>