rework region management in the topo
The previous implementation of the cell to region mapping
relied on lazily populating the mapping of cells to regions as
part of the query routing tier.

Rework the API so that the region mapping is initialized
proactively as part of the gateway initialization path, which
allows the map to be used without a wrapping mutex by the
query serving tier and better reflects the intended design.

Signed-off-by: Michael Demmer <mdemmer@slack-corp.com>